Accidentally work on server testing within webpack branch
