+++ db/index.js
@@ -30,4 +30,5 @@ const itemSchema = mongoose.Schema({
 let Item = mongoose.model('Item', itemSchema);
 
 
-module.exports = Item;
+module.exports.Item = Item;
+module.exports.mongoose = mongoose;
