+++ package.json
@@ -23,8 +23,11 @@
     "@babel/core": "^7.11.6",
     "@babel/preset-env": "^7.11.5",
     "@babel/preset-react": "^7.10.4",
+    "@testing-library/react": "^11.0.4",
     "babel-loader": "^8.1.0",
     "jest": "^26.4.2",
+    "superagent": "^6.1.0",
+    "supertest": "^5.0.0",
     "webpack": "^4.44.2",
     "webpack-cli": "^3.3.12"
   }
